OpenGL: Revive and resurrect ANGLE support
The removal of ANGLE causes Qt apps that need HDR support to implement a
custom solution for each and every single supported graphics API, as
Qt6's Rhi doesn't support high-bit-depth rendering or non-sRGB surfaces.

This commit brings back the support that was removed in Qt 6.0, makes
that code compliant with the new platform abstraction API, makes the
code consistent with the new C++17 standard applicable to 6.0, and updates
FindEGL to 5.107.0-rc1 to allow it to find Krita-built EGL on Windows
(which correctly uses the lib prefix).

This also adds some logging code to hellogles3 to verify that libEGL
loads successfully and prints the ANGLE renderer information.

This reverts commit 752497910b67b2a1a80560840ca44548d8893434.

This reverts commit 8158a8767c4c0ece37c55c2cc6777267221693ee.

Change-Id: Idf0e43c501c128f3cf48ac0ceb1edc760873b942