QSurfaceFormat: implement color space selection for OpenGL ES
With the patch one can select the color space of the OpenGL ES
surface which is used as root surface of the underlying native
window.

This feature is needed, e.g. when the user wants to render HDR
content on screen. In such a case OS should be instructed about
how to treat the graphical data in the application framebuffer.

The easiest approach is to call QSurfaceFormat::setDefaultFormat()
before creating the first application window. In such a case the
root surface will (may) be in the requested format.

Supported color spaces/formats:

1) sRGB, SDR
2) scRGB (Rec 709, gamma 1.0), HDR
3) Rec 2020 PQ, HDR

Please take into account that in real life the user should select
the proper bit depth for each color space, otherwise the system
will refuse to create the surface:

1) sRGB --- 8 bit or 10 bit
2) scRGB --- 16 bit only
3) Rec 2020 PQ --- 10 bit only

For scRGB, this is done automatically by the surface configuration
chooser.

Please note that, on Windows, color space selection is supported only
on platforms with DXGI 1.4 and higher.

See:

- https://invent.kde.org/szaman/qtbase/-/commit/9a8fbae77eb836d6ac1ae385b63e0c399fbe8eb5
- https://invent.kde.org/szaman/qtbase/-/commit/2cc62140938972cb7e6845bbeaa8f9a64e8ca04b
- https://invent.kde.org/szaman/qtbase/-/commit/f5ae313a070a1a6da92db88b27dfbd1e6366ffa9

Co-authored-by: Dmitry Kazakov <dimula73@gmail.com>
Co-authored-by: Alvin Wong <alvin@alvinhc.com>
Change-Id: I8c8cabc31f8ab545ab4f17cd1ef33c3cada13f5a